examen verwijderen

+++ controllers/ExamenController.php
@@ -4,6 +4,7 @@ namespace app\controllers;
 
 use Yii;
 use app\models\Examen;
+use app\models\Form;
 use app\models\ExamenSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
@@ -117,8 +118,13 @@ class ExamenController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
-
+        if ( count(form::find()->where(['examenid'=>$id])->all()) == 0 ) {
+            $this->findModel($id)->delete();
+            Yii::$app->session->setFlash('success', "Exam deleted.");
+        } else {
+            Yii::$app->session->setFlash('error', "Exam cannot be deleten, it has still forms attached to it.");
+        }
+       
         return $this->redirect(['index']);
     }
 
@@ -144,5 +150,55 @@ class ExamenController extends Controller
         $params = array(':id'=> $id);
         Yii::$app->db->createCommand($sql)->bindValues($params)->execute();
         return $this->redirect(['index']);
+    }
+
+    public function actionCopyExam($id) {
+        
+        $fromExamenId=$id;
+        // get examen
+        $examen=Examen::find()->where(['id'=>$id])->one();
+    
+        if (! $examen) {
+            return;
+        }
+        // copy exam
+        $sql="  insert into examen (naam, actief, datum_start, datum_eind, examen_type, otherid, titel)
+                select concat(naam, ' copy'), actief, datum_start, datum_eind, examen_type, otherid, titel
+                from examen where id=:id";
+        $params = [':id'=> $fromExamenId];
+        $result=Yii::$app->db->createCommand($sql)->bindValues($params)->execute();
+
+        $sql="  select max(id) id from examen";
+        $result=Yii::$app->db->createCommand($sql)->queryOne();
+        $toExamenId=$result['id'];
+
+        //$toExamenId = Yii::$app->db->getLastInsertID();
+        //$toExamenId = 9;
+
+        $sql="  select f.id from form f inner join examen e on e.id=f.examenid where e.id=:fromExamenId";
+        $params = [':fromExamenId'=> $fromExamenId];
+        $forms=Yii::$app->db->createCommand($sql)->bindValues($params)->queryAll();
+
+        // itterate through forms
+        foreach($forms as $fromForm) {
+            d($fromForm['id']);
+            $sql="  insert into form (omschrijving, nr, examenid, actief, werkproces, instructie)
+                    select concat(f.omschrijving, ' copy') , f.nr, :toExamenId, f.actief, f.werkproces, f.instructie
+                    from form f
+                    where id=:formId";
+            $params = [':formId'=> $fromForm['id'], ':toExamenId'=> $toExamenId ];
+            $result=Yii::$app->db->createCommand($sql)->bindValues($params)->execute();
+            $toFormId = Yii::$app->db->getLastInsertID();
+
+            $sql="  insert INTO vraag (formid, volgnr, vraag, toelichting, ja, soms,nee, mappingid, standaardwaarde)
+                    select :toFormId, volgnr, vraag, toelichting, ja, soms,nee, mappingid, standaardwaarde FROM `vraag` 
+                    where formid = :fromFormId";
+            $params = [':fromFormId'=> $fromForm['id'], ':toFormId'=> $toFormId ];
+            $result=Yii::$app->db->createCommand($sql)->bindValues($params)->execute();       
+
+        }
+        
+        return $this->redirect(['index']);
+
     }
 }
